sorting npm stuff out
